+++ src/window_navigator2.h
@@ -1,0 +1,33 @@
+/**
+ * File:   window_navigator2.h
+ * Author: AWTK Develop Team
+ * Brief:  window_navigator2 source
+ *
+ * Copyright (c) 2018 - 2021  Guangzhou ZHIYUAN Electronics Co.,Ltd.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * License file for more details.
+ *
+ */
+
+/**
+ * History:
+ * ================================================================
+ * 2021-05-11 Li XianJing <xianjimli@hotmail.com> created
+ *
+ */
+
+#ifndef WINDOW_NAVIGATOR2_H
+#define WINDOW_NAVIGATOR2_H
+
+#include "awtk.h"
+
+BEGIN_C_DECLS
+
+ret_t window_navigator2_open(widget_t* to_close);
+
+END_C_DECLS
+
+#endif /*WINDOW_NAVIGATOR2_H*/